requirements.txt
aiogram
aiohttp
aiosend
aiosqlite
certifi
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dateutil
redis
requests
yookassa
python-dotenv